requirements.txt
ipython==6.5.0
natasha==0.10.0
pytest==3.8.0
pytest-runner==4.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-telegram-bot==11.1.0
pytz==2018.5
